breaking up initialization to it's own file, improving readability and troubleshooting
